Optimize validate_sequence_{any,all} via min max instead of sum.
To assess validity of all values, can look at min: when it's 1, it means all elements are 1.

The same goes for any value being valid, can look at max: when it's 1, it means there's an element being 1.

This simplifies calculation further. Summing is more heavy in comparison, so this is another significant improvement for this part.