Avoid compute() in scaler.
This does require exporting the zero values check, but through the scale() call, as well as save(). Then the functions get called by dask later.

Next step is to avoid compute() in _create_sample_index, so the only compute is the one in forecast_basedataset once.

Otherwise, the many lazy operations get recomputed.