Setup learning rate epochs drop to 0.
This is the amount of epocs to hold off LR reduction. When 1, it means that e.g. given two epocs where loss didn't decrease (along with a threshold by default) or increase, an additional 1 epoc is made prior to reducing LR.

Loss should be reduced each epoc. So this value may be 0 instead to avoid wasteful epocs when LR should be stepped down.