Address UnstableSpecificationWarning: use variable strings.
This converts old types of Unicode or Byte String data to the variable length string type, which is better supported in zarr.

While it's possible to use the most modern numpy.dtype.StringDType, it'd require setting up explicitly an encoding dict for the to_zarr call, akin to `encoding = { name: {"dtype": "str"} for name, coord in ds.coords.items() if isinstance(coord.dtype, np.dtypes.StringDType)}`, but it's more complicated. So it's simpler to let the framework assign these automatically.