Detach scaler check zero and save from the graph.
Dask needs to serialize the parent subgraph upwards to send args to Delayed function, even if the leaf node is very small.

So, we need to revert to the design where the validation itself has to be not part of the graph. Runtime now is immediate even for 3000 basins over 40yrs.

Another option may be to make a division by zero inherent error when a scale is zero, but it won't be descriptive and confusing.