Setup CUDA 12.4 and explicitly setup versions.
1. Cuda 11.8 has been deprecated.
2. The driver NVIDIA-SMI version 550 that supports 11.8 supports 12.4 (`$ nvidia-smi`).
3. 11.8 results in using old deprecated libraries (numpy 1.x instead of 2.x, dask, xarray, etc).

So, upgrading and explicitly setting everything to comply with 12.4 is ok both in compatibility and correctness (version consistency).

Otherwise, running the various tools on the terminal to assess the cuda version, we get different versions. We use pytorch for cuda 11.8 but we use (the installed version) cuda 12.9-3 yet versions from CLIs report 11.8 and 12.4 being used also at the same time.